Updated My Task List to fully implement Core Data
Add/Remove/Load all talk to Core Data now. Furthermore, Core Data was implemented fully in the TaskManager object, so that the view controller doesn't have to be directly aware of the model.

It's not done the most efficient way possible. After any change is made (add/remove), it reloads from the database. However, it keeps the code simple. The results are stored in an array, and the getTask() and getTaskCount() methods just reference the entities in the array.